requirements.txt
absl-py==0.15.0
asgiref==3.7.2
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
dj-database-url==2.0.0
Django==3.2
django-heroku==0.3.1
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
idna==3.4
importlib-metadata==6.6.0
jax==0.4.11
keras==2.12.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
ml-dtypes==0.2.0
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
protobuf==3.20.3
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygame==2.4.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.15.0
sqlparse==0.4.4
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==1.1.0
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.3.5
whitenoise==6.5.0
wrapt==1.12.1
zipp==3.15.0
